Correcion de float a decimal, float no es preciso para el calculo de dinero, decimal es mas preciso y recomendado para este tipo de calculos.(me confundi con python)

+++ entrega1.cs
@@ -6,15 +6,15 @@ class enviodepaquetes
     // el estatic void es el punto de entrada del programa, es donde se ejecuta el codigo y se llama a las funciones o metodos que se quieran usar, es como el motor del programa y por si solo no funciona y funciona sin objetos
     static void Main()
     {
-        float monto_pedido = 0;
+        decimal monto_pedido = 0;
         string ciudad_destino = "";
         bool tipo_cliente;
         int cantidad_items = 0;
         string categoria_despacho ="";
-        float costo_envio = 0;
+        decimal costo_envio = 0;
   //el parse se usa simpre ya que el computador entrega un string entonces se parsea a un numero para poder hacer operaciones matematicas con el :V
     Console.WriteLine("Ingrese el monto del pedido:");
-    monto_pedido = float.Parse(Console.ReadLine());
+    monto_pedido = decimal.Parse(Console.ReadLine());
 
   Console.WriteLine("¿El envio es internacional? (si/no)");
 string respuesta = Console.ReadLine();
